+++ SandboxiePlus/SandMan/Engine/BoxEngine.cpp
@@ -8,10 +8,19 @@
 #include "SysObject.h"
 #include "WizardObject.h"
 
+#include "ScriptManager.h"
+
+#include "../MiscHelpers/Common/Common.h"
+
+#include "../SandMan.h"
+#include "../Views/SbieView.h"
+#include "../Windows/OptionsWindow.h"
+#include "../Windows/SettingsWindow.h"
 #include "../Wizards/BoxAssistant.h"
 
 #include <private/qv4engine_p.h>
 #include <private/qv4script_p.h>
+#include <private/qv4qobjectwrapper_p.h>
 
 int CBoxEngine::m_InstanceCount = 0;
 
@@ -56,27 +65,11 @@ QV4::ReturnedValue method_translate(const QV4::FunctionObject *b, const QV4::Val
 
     QString Result;
     for (int i = 0; i < argc; i++) {
-        if (i == 0) Result = (Wizard ? Wizard->Tr(argv[i].toQStringNoThrow()) : argv[i].toQStringNoThrow());
+        if (i == 0) Result = theGUI->GetScripts()->Tr(argv[i].toQStringNoThrow());
         else Result = Result.arg(argv[i].toQStringNoThrow());
     }
 
     return QV4::Encode(scope.engine->newString(Result));
-}
-
-QV4::ReturnedValue method_print(const QV4::FunctionObject *b, const QV4::Value *v, const QV4::Value *argv, int argc)
-{
-    QV4::Scope scope(b);
-    QV4::ExecutionEngine *v4 = scope.engine;
-
-    QString Result;
-    for (int i = 0; i < argc; i++) {
-        if (i > 0) Result.append(" ");
-        Result.append(argv[i].toQStringNoThrow());
-    }
-    CBoxEngine* pEngine = qobject_cast<CBoxEngine*>(CJSEngineExt::getEngineByHandle(v4)->thread());
-    pEngine->AppendLog(Result);
-
-    return QV4::Encode::undefined();
 }
 
 QV4::ReturnedValue method_wcmp(const QV4::FunctionObject *b, const QV4::Value *v, const QV4::Value *argv, int argc)
@@ -90,24 +83,58 @@ QV4::ReturnedValue method_wcmp(const QV4::FunctionObject *b, const QV4::Value *v
     return QV4::Encode(CSbieUtils::WildCompare(argv[0].toQStringNoThrow(), argv[1].toQStringNoThrow()));
 }
 
+QV4::ReturnedValue method_invoke(const QV4::FunctionObject *b, const QV4::Value *v, const QV4::Value *argv, int argc)
+{
+    QV4::Scope scope(b);
+    QV4::ExecutionEngine *v4 = b->engine();
+
+    QString Name = argv[0].toQStringNoThrow();
+
+    CBoxEngine* pEngine = qobject_cast<CBoxEngine*>(CJSEngineExt::getEngineByHandle(v4)->thread());
+    CJSEngineExt* pJSEngine = pEngine->GetEngine();
+
+    QString Script;
+    QMetaObject::invokeMethod(theGUI->GetScripts(), "GetScript", Qt::BlockingQueuedConnection, Q_RETURN_ARG(QString, Script), Q_ARG(QString, Name));
+    if (!Script.isEmpty()) {
+        QJSValue ret = pJSEngine->evaluateScript(Script, Name + ".js");
+        if (ret.isError()) {
+            pJSEngine->throwError(ret);
+            return QV4::Encode::undefined();
+        } else {
+            QV4::ScopedValue rv(scope, scope.engine->fromVariant(ret.toVariant()));
+            QV4::Scoped<QV4::QObjectWrapper> qobjectWrapper(scope, rv);
+            if (!!qobjectWrapper) {
+                if (QObject *object = qobjectWrapper->object())
+                    QQmlData::get(object, true)->setImplicitDestructible();
+            }
+            return rv->asReturnedValue();
+        }
+    }
+
+    return QV4::Encode(false);
+}
+
 void CBoxEngine::init()
 {
+    connect(m_pEngine, SIGNAL(printTrace(const QString&)), this, SLOT(AppendLog(const QString&)));
+
     QV4::Scope scope(m_pEngine->handle());
     scope.engine->globalObject->defineDefaultProperty(QStringLiteral("tr"), method_translate);
-    scope.engine->globalObject->defineDefaultProperty(QStringLiteral("print"), method_print);
     scope.engine->globalObject->defineDefaultProperty(QStringLiteral("wildCompare"), method_wcmp);
+    scope.engine->globalObject->defineDefaultProperty(QStringLiteral("invoke"), method_invoke);
 
     m_pEngine->globalObject().setProperty("system", m_pEngine->newQObject(new JSysObject(this)));
     m_pEngine->globalObject().setProperty("sbie", m_pEngine->newQObject(new JSbieObject(new CSbieObject(this), this)));
 }
 
-bool CBoxEngine::RunScript(const QString& Script, const QString& Name)
+bool CBoxEngine::RunScript(const QString& Script, const QString& Name, const QVariantMap& Params)
 {
     if (isRunning())
         return false;
 
     m_Script = Script;
     m_Name = Name;
+    m_Params = Params;
     m_State = eRunning;
 
     if(!m_pEngine) m_pEngine = new CJSEngineExt(); // the engine lives in its own thread
@@ -143,6 +170,9 @@ void CBoxEngine::run()
     //QElapsedTimer timer;
     //timer.start();
     
+    for (auto I = m_Params.begin(); I != m_Params.end(); ++I)
+        m_pEngine->globalObject().setProperty(I.key(), m_pEngine->toScriptValue(*I));
+
     //auto ret = m_pEngine->evaluateScript("(()=>{" + m_Script + "})()", m_Name);
     auto ret = m_pEngine->evaluateScript(m_Script, m_Name);
 
@@ -215,7 +245,11 @@ bool CBoxEngine::WaitLocked() {
 
     // WARNING: call this function only from the engine thread itself !!!
 
-    m_Wait.wait(&m_Mutex);
+    if (m_pDebuggerBackend) {
+        while (!m_Wait.wait(&m_Mutex, 10))
+            QCoreApplication::processEvents(); // for CV4DebugAgent::runJob()
+    } else
+        m_Wait.wait(&m_Mutex);
 
     emit StateChanged(m_State);
 
@@ -324,8 +358,12 @@ void CWizardEngine::init()
 
 void CWizardEngine::SetState(EState state, const QString& Text)
 {
-    if (state == eError)
-        m_Report["Error"] = Text;
+    if (!Text.isEmpty()) {
+        if (state == eError)
+            m_Report["Error"] = Text;
+        else if (state == eFailed)
+            m_Report["Failure"] = Text;
+    }
     CBoxEngine::SetState(state, Text);
 }
 
@@ -383,4 +421,34 @@ bool CWizardEngine::IsNoAppliedShadow(const QString& OriginalName)
     if (I != m_Shadows.end())
         return I->iApplyChanges == 0;
     return false;
+}
+
+void CWizardEngine::OpenSettings(const QString& page)
+{
+    SetState(eReady);
+
+    CSettingsWindow* pSettingsWnd = new CSettingsWindow();
+	connect(pSettingsWnd, &CSettingsWindow::Closed, this, [=]() {
+		Continue(false);
+	});
+    //theGUI->SafeExec(pSettingsWnd);
+    pSettingsWnd->showTab(page);
+}
+
+void CWizardEngine::OpenOptions(const QString& box, const QString& page)
+{
+    SetState(eReady);
+
+    CSandBoxPtr pBox = theAPI->GetBoxByName(box);
+    if (pBox.isNull()) {
+        Continue(false);
+        return;
+    }
+    
+    COptionsWindow* pOptionsWnd = new COptionsWindow(pBox, pBox->GetName());
+	connect(pOptionsWnd, &COptionsWindow::Closed, this, [=]() {
+		Continue(false);
+	});
+    //theGUI->SafeExec(pOptionsWnd);
+    pOptionsWnd->showTab(page);
 }