+++ SandboxiePlus/SandMan/Troubleshooting/Sandboxing/SBIEMSG/SBIE2113.js
@@ -1,0 +1,13 @@
+/*
+*	group: other
+*	class: sandboxing
+* name: SBIE2113 File is too large to copy into sandbox, creating empty file
+* description: File is too large to copy into sandbox, creating empty file
+*
+*/
+
+if (typeof msgCode === 'undefined') {
+    msgCode = 2113;
+}
+
+invoke("SBIECOPY");