+++ SandboxiePlus/SandMan/Troubleshooting/Sandboxing/SBIEMSG/SBIE2114.js
@@ -1,0 +1,13 @@
+/*
+*	group: other
+*	class: sandboxing
+* name: SBIE2114 File is too large to copy into sandbox, denying access
+* description: File is too large to copy into sandbox, denying access
+*
+*/
+
+if (typeof msgCode === 'undefined') {
+    msgCode = 2114;
+}
+
+invoke("SBIECOPY");